+++ src/main/java/domain/path/PathInfo.java
@@ -1,4 +1,7 @@
-package domain;
+package domain.path;
+
+import domain.board.Position;
+import domain.piece.Piece;
 
 public record PathInfo(Position position, Piece piece) {
     public boolean hasPiece(){
