+++ src/main/java/domain/piece/Camp.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.piece;
 
 public enum Camp {
     CHO,
