+++ src/main/java/domain/piece/Piece.java
@@ -1,4 +1,7 @@
-package domain;
+package domain.piece;
+
+import domain.path.PathInfo;
+import domain.board.Position;
 
 import java.util.List;
