+++ src/test/java/domain/board/BoardTest.java
@@ -1,5 +1,8 @@
-package domain;
+package domain.board;
 
+import domain.piece.Camp;
+import domain.piece.Piece;
+import domain.piece.PieceType;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
