+++ src/test/java/domain/piece/DirectionTest.java
@@ -1,5 +1,6 @@
-package domain;
+package domain.piece;
 
+import domain.board.Position;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
